reconfigure for c/8 longshot
NG, filled at least 20% of memory in opening expand.  Need to redo to
make nodes (B, B, usize) indicating number of bits of second B that have
been built so far...

Can build more than one bit in an expand if it comes to it, but at a
minimum we cannot afford to fill memory with the full first row expand.